Create kernel events by writing to proc file
This feature allow to write events in the kernel trace stream by writing to the
file /proc/lttng_user_event. The content is written unmodified to the trace as
string type. The maximum string length saved per event is 256 bytes. The event
type and the proc file are accessibles when the module lttng-user-event is
loaded.

This is a prototype implementation. The final implementation should avoid the
temp copy of the string on the kernel stack.