Expose kernel tracer to user-space (version 7)
By writing to the file /proc/lttng_uevent, a user-space application creates a
kernel event. The event's payload is by default UTF-8 text, but any data can be
written. The maximum size of the data is at most the size of a packet defined
for the trace session. Null-character is optional and is not enforced.  The
event uses sequence for space efficiency and to store any data as payload.

The feature is enabled when the module lttng-uevent is loaded. The module can
be removed if no tracing sessions with lttng_uevent enabled is active.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>